# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aegis Platform API
  version: v1
  description: >
    HTTP API for the Aegis enterprise access-control platform — a multi-tenant,
    microservices SaaS substrate (centralized RBAC + ABAC PDP/PEP, PostgreSQL
    row-level tenant isolation, signed service-to-service calls, hash-chained
    audit). Seven business services sit behind a single gateway.


    **Edge routing.** All authenticated traffic enters through the gateway, which
    resolves the first path segment (`user-management`, `expense`, `payroll`,
    `reporting`, `workflow`, `notification`, `invoice`) to a downstream service
    and forwards the request unchanged. Each service independently re-validates
    auth via its own PEP (defense in depth). Every external path therefore has the
    form `/{service}/v1/...` (public prefix `/v1`). `/health` on each service and
    on the gateway bypasses auth/context middleware.


    **Required context headers.** Authenticated requests must carry
    `X-Tenant-Id` and `X-Correlation-Id` (the gateway mints the correlation id at
    the edge if absent and echoes it on every response, including error hops).
    `X-Caller` is set to `gateway` on every forwarded hop. The tenant is ambient
    (driven into PostgreSQL RLS via `SET LOCAL app.current_tenant`) — there is no
    tenant path segment. Money/state-changing writes may honor an
    `Idempotency-Key` header (pay-run disburse).


...
    the principal also holds `payroll.sensitive.read`.


...
  contact:
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Authenticated user JWT issued by user-management `/auth/login`. Carried as
        `Authorization: Bearer <token>`. The gateway propagates it downstream; each
        service validates it and resolves the principal's permissions.
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      description: Tenant (organization) the request operates within. Drives PostgreSQL RLS.
      schema:
        type: string
    CorrelationHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        Request-tracking id. Minted at the gateway if absent and echoed on every
        response (including error hops). Required on every internal hop.
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorEnvelope:
      type: object
      description: Terminal error shape emitted by the service-core error middleware (and the gateway).
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: Stable machine code, e.g. E_VALIDATION, E_UNAUTHORIZED, E_FORBIDDEN, E_NOT_FOUND, E_CONFLICT, E_GATEWAY_TIMEOUT.
              type:
                type: string
                description: Error category, e.g. VALIDATION, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, GATEWAY, SYSTEM.
              message:
                type: string
              correlationId:
                type: string
            required: [code, type, message, correlationId]
      required: [errors]
...
    SetConfigRequest:
      type: object
      properties:
        value:
          description: Arbitrary config value (any JSON type).
      required: [value]
    SetFlagRequest:
      type: object
      properties:
        enabled: { type: boolean }
      required: [enabled]

    # ---- expense ----
...
  responses:
    BadRequest:
      description: Validation error (E_VALIDATION).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing/invalid token or required headers (E_UNAUTHORIZED).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but lacks the required permission, or ABAC policy denied (E_FORBIDDEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ServerError:
      description: Unexpected server error (E_SYSTEM / E_DATABASE).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
paths:
...
  /user-management/v1/tenant/config:
    get:
      tags: [user-management]
      summary: List tenant config entries
      description: Requires permission `tenant.view`.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      responses:
        '200':
          description: Tenant config entries.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /user-management/v1/tenant/config/{key}:
    put:
      tags: [user-management]
      summary: Set a tenant config value
      description: Requires permission `tenant.manage`.
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetConfigRequest' }
      responses:
        '200':
          description: Config value set.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /user-management/v1/tenant/features:
    get:
      tags: [user-management]
      summary: List feature flags
      description: Requires permission `tenant.view`.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      responses:
        '200':
          description: Feature flags.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /user-management/v1/tenant/features/{flag}:
    put:
      tags: [user-management]
      summary: Toggle a feature flag
      description: Requires permission `tenant.manage`.
      parameters:
        - name: flag
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetFlagRequest' }
      responses:
        '200':
          description: Flag set.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ===================== EXPENSE =====================
...